pyproject.toml
[project]
name = "speech-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deep-translator>=1.11.4",
    "python-docx>=1.1.2",
    "speechrecognition>=3.14.2",
    "streamlit>=1.44.0",
]
